+++ src/constants/index.ts
@@ -15,3 +15,4 @@ export * from './openai.const';
 export * from './premium.const';
 export * from './defaultIcons';
 export * from './contextMenu.const';
+export * from './table.const';