+++ src/constants/table.const.ts
@@ -1,0 +1,7 @@
+export const TABLE_CLASS_NAME = 'Ink-table';
+export const TABLE_TAG = 'table';
+export const TABLE_ROW_TAG = 'tr';
+export const TABLE_SECTION_THEAD = 'thead';
+export const TABLE_CELL_TAG_TD = 'td';
+export const TABLE_CELL_TAG_TH = 'th';
+export const TABLE_CELL_SELECTOR = 'td, th';